✅ test(pingpong): refactor integration tests for clarity and cancellation
- refactor ping-pong integration tests for better readability and structure
- move stream crossover helpers to a shared test infrastructure file
- improve cancellation token propagation in tests and services
- expose pong service method to simplify direct testing scenarios
